Reworked shops to give predesigned pets
- This rework is important for when I give the pets specific models, they can't just be randomly generated but instead picked from pools of premade ones
- This will allow, later on, the ability to combine multiple of the same pets into stronger versions of them
- Pet rarity now works as a weight out of 100, essentially what the percentage chance of getting them is, rather than the cheap and dirty method of "roll twice with disadvantage" that I was doing before
- Pet registry has been created, with each shop having an index and 5 pets in their registry

+++ Assets/Scripts/Pet.cs
@@ -4,10 +4,10 @@ public class PetDetails
 {
     public string Name;
     public float Strength = 1f;
-    public int Rarity = 1;
+    public float Rarity = 1;
     public string Material;
 
-    public PetDetails(string name, float strength, int rarity, string material) 
+    public PetDetails(string name, float strength, float rarity, string material) 
     {
         Name = name;
         Strength = strength;
